Move bitwise.rs
